code cleanup

+++ java/code/src/com/redhat/rhn/domain/scc/SCCRepository.java
@@ -275,7 +275,7 @@ public class SCCRepository extends BaseDomainHelper {
                     return Optional.of(a);
                 }
             }
-            if (result.isEmpty() || a.getId().longValue() < result.get().getId().longValue()) {
+            if (result.isEmpty() || a.getId() < result.get().getId()) {
                 // get always the same result and use the oldest alternative if there are multiple
                 result = Optional.of(a);
             }
